+++ ormar/__init__.py
@@ -76,7 +76,7 @@ class UndefinedType:  # pragma no cover
 
 Undefined = UndefinedType()
 
-__version__ = "0.10.10"
+__version__ = "0.10.11"
 __all__ = [
     "Integer",
     "BigInteger",